Notebook support to new SimulationApplication
Minor review comment fixes
1. /common was symbolic link. Kept it the same and made modification to the ../common (original file)
2. Removed all the commented code in the Dockerfile

Testing:
Tested it murasuni account, science-team account and burner account. It works.
